Add Project 2
Circuit diagram as in book.

This is a nice little test program that introduces interactivity and
runs at human speeds.

As a deviation from the book’s code I’ve put the second sleep outside
of the loop to reduce the rate of polling when in the rest state. In
practice this doesn’t make much of a difference, but there’s no need
for the additional responsiveness here.

Note to self: Don’t set your inputs to be outputs and your outputs to
be inputs. It doesn’t work so well.