Make a Light Theremin
Circuit diagram as in book.

This was fun. Make a bleepy thing. The program uses sensor data to
calibrate during the first five seconds of running and then uses that
range for pitch detection.

This is fine, but there’s much that could be done. It’s nice to see
that something so simple can achieve something remotely musical.
However, a real Theremin would want a sine wave, so a DAC is on the
list of things to look up.

More over, the calibration time is not ideal. Using the reset button is
fine, but five seconds is a long time and it mutes the program. A
second button could be useful here.

Thirdly, the tone() function does not have an amplitude argument (which
makes sense as it’s a digital write) and we have no other options for
this here. Wiring a second photo-resistor with the buzzer works, but
isn’t very refined.

More to be done.