Add Iterations on DAC Oscillator
Project 01 and 02 document the proof of concept code. Project 03
produces a stable output at a reasonable frequency range.

What follows is the code that drives a R2R resistor bank to create a
DAC. It seems that the serial communication between the Arduino and the
computer skews the timing too much to give a usable output.

To avoid the overhead of calculating the output value on the fly a
wavetable is pre-populated. To avoid the complexity of interpolation
Project 03 uses a variable timer to cycle through the wavetable at the
required frequency. (This might cause issues moving to a more complex
sound generation model.)

TODO: Have a variable oscillator frequency and input sensing. (Project
03 gives a usable output for a reasonable range, but no way of changing
it dynamically.)