Add Antenna Proof of Concept
Based on the CapativeSensor library by Paul Badger
http://playground.arduino.cc/Main/CapacitiveSensor

The sensor library uses a pair of digital pins to determine the
relative capacitance of the intermediate circuit by measuring the time
taken for the read pin to measure a high state.

This should be fine for our purposes. The reading time is variable
(based on the capacitance) and the return value is arbitrary, but there
is enough to work with that some manipulation of the results should
work.