Add Capacitive Sensing to Theremin Sketch
Integrating the CapacitiveSensor test code for pitch control to the
main body of code.

A 10M Ohm resistor and an antenna gives a good range of input values.
However, there’s limited stability to the values, so some smoothing of
the value changes has been added. There is also some jittering to the
baseline value readings that can be looked into further.