Add sample averaging
Done some debugging and testing to try and improve the stability.
However, the limits of the Arduino are being pushed with the DAC. The
constant changing of PORTD at variable frequencies is giving an
inconsistent baseline for capacitance readings. Keeping the pitch fixed
have a constant reading.

There may be ways around this, but I don’t have the components to deal
with it before the weekend, so I will look into a hybrid model that has
a stable pitch with a photo diode and a capacitive volume control (that
doesn’t need the same precision).

Long term I may look at replacing the DAC with a VCO, which will have a
smaller variation in load (using a single pin for PMW), but for now
this will do.